+++ app.py
@@ -14,6 +14,7 @@ pagina = st.navigation(
     [
         st.Page("views/build_your_menu.py", title="Build your menu", icon="🥗"),
         st.Page("views/editor_recetas.py", title="Editor de recetas", icon="🧑‍🍳"),
+        st.Page("views/editor_ingredientes.py", title="Editor de ingredientes", icon="🥕"),
         st.Page("views/personas.py", title="Personas", icon="🧑‍🤝‍🧑"),
     ]
 )